LayoutTests/http/tests/security/cross-origin-plugin-allowed.html

<html>
<head>
<script>
if (window.testRunner) {
	testRunner.dumpAsText();
	testRunner.dumpChildFramesAsText();
}

</script>
<body>
<p>This iframe should not have private browsing enabled:</p>
<iframe src="http://localhost:8000/security/resources/cross-origin-iframe-for-plugin.html"></iframe>
<p>This iframe should not have private browsing enabled:</p>
<iframe src="http://127.0.0.1:8000/security/resources/cross-origin-iframe-for-plugin.html"></iframe>
</body>
</html>
