LayoutTests/http/tests/security/cross-origin-plugin-allowed-expected.txt

This iframe should not have private browsing enabled:


This iframe should not have private browsing enabled:



--------
Frame: '<!--frame1-->'
--------
false

--------
Frame: '<!--frame2-->'
--------
false
